Moved Preview Pane and Context Tree Sidebar to their own files, out from Notebook file. Other misc cleanup and removal of unused imports and comments
