+++ beaker-vue/src/components/BeakerContextSelection.vue
@@ -40,7 +40,7 @@
 
 <script setup lang="ts">
 
-import { defineProps, defineEmits, ref, defineModel, onMounted, computed, nextTick, onBeforeMount, watchEffect } from "vue";
+import { defineProps, defineEmits, ref, onMounted, computed, watchEffect } from "vue";
 import Dialog from 'primevue/dialog';
 import Button from 'primevue/button';
 import InputGroup from 'primevue/inputgroup';
